Added `DescendRange`, `DescendLessOrEqual`, `DescendGreaterThan`, and `Descend` functions. These are modeled after the existing `Ascend..` methods.

Also modified the iteration function to remove the `from()` and `to()` in favor of adding a `start` and `stop` parameters instead. There is a modest performance ~10% boost by avoiding superfluous function calls.

before:
```
BenchmarkAscend-8     	   10000	    216529 ns/op
BenchmarkAscendRange-8	    3000	    433896 ns/op
```  
after:
```
BenchmarkAscend-8     	   10000	    196143 ns/op
BenchmarkAscendRange-8	    3000	    396469 ns/op
```

The benchmark code is not included in the source, but I did create a gist at [iteration_test.go](https://gist.github.com/tidwall/378976e24be7e2cb5a74b10eeba744eb)

I hope you find this update useful and thanks a ton for such a fantastic library!