@gconnell 2016-08-29
Ping this once the above changes (benchmarks, coment changes on reverse ranges) are in place, and I think we should be good to go.