@tidwall 2016-09-23
Ping

I made the changes a few weeks back, but forgot to ping like you asked. Sorry.