@gconnell 2016-09-23
Merged, thanks!